refactor(teams-chat-sync): extract shared Mongo store into pkg/teamsstore
Moves the teams_user/teams_chat Mongo layer (collection wiring, projections,
$setOnInsert/$set upsert semantics) into pkg/teamsstore — same pattern as
pkg/userstore — so downstream consumers (e.g. a member-sync job reading
needMemberSync) reuse it. Models remain in pkg/model; teams-chat-sync keeps
its consumer-side interfaces, now satisfied by *teamsstore.Store.
