+++ pkg/teamsstore/teamsstore_test.go
@@ -1,4 +1,4 @@
-package main
+package teamsstore
 
 import (
 	"testing"
