+++ teams-chat-sync/store.go
@@ -11,7 +11,7 @@ import (
 //go:generate mockgen -source=store.go -destination=mock_store_test.go -package=main
 
 // TeamsUserStore reads the externally-populated teams_user collection and
-// advances per-user watermarks.
+// advances per-user watermarks. Satisfied by *teamsstore.Store.
 type TeamsUserStore interface {
 	// ListUsers returns every teams_user projected to exactly the fields the
 	// sync needs (_id, siteID, account, from).
@@ -23,7 +23,7 @@ type TeamsUserStore interface {
 // TeamsChatStore upserts synced chats keyed on _id. oneOnOne chats are
 // insert-only; for other chat types createdDateTime and siteID are
 // $setOnInsert-only while the mutable fields (including each chat's
-// build-time UpdatedAt stamp) are refreshed.
+// build-time UpdatedAt stamp) are refreshed. Satisfied by *teamsstore.Store.
 type TeamsChatStore interface {
 	UpsertChats(ctx context.Context, chats []model.TeamsChat) error
 }
